backend/requirements.txt
django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
requests==2.32.4
django-environ==0.11.2
